feat: also show invoice info on thank you / my-account order view
Extracts the row-building logic into a reusable InvoiceInfo::get_rows()
method and hooks into woocommerce_order_details_after_customer_address
so the same invoice fields are rendered inside the billing <address>
block of the order-received (thank you) page and the my-account order
view — mirroring the email behavior.

Renames EmailInvoiceDetails → InvoiceInfo and renames the filter
hezarfen_email_billing_invoice_rows → hezarfen_invoice_info_rows to
reflect the broader scope.
